fix(indexing): show lang item on open files
The active editor will display the language item for the hardhat config
once indexing has passed it.
